+++ .idea/workspace.xml
@@ -6,8 +6,15 @@
   <component name="ChangeListManager">
     <list default="true" id="0c2d425c-1943-4e9c-8fd7-b83aff504fc7" name="Changes" comment="https://github.com/MilianCode/banking-system.git">
       <change beforePath="$PROJECT_DIR$/.idea/dataSources/1f4d7f95-301a-489c-857e-c8747c073cea.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/dataSources/1f4d7f95-301a-489c-857e-c8747c073cea.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/inspectionProfiles/Project_Default.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/inspectionProfiles/Project_Default.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/customer/Customer.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/customer/Customer.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/customer/CustomerBalanceManaging.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/customer/CustomerBalanceManaging.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/customer/CustomerTransactionManaging.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/customer/CustomerTransactionManaging.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/MainWork.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/MainWork.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/transaction/Transaction.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/transaction/Transaction.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/target/classes/customer/CustomerTransactionManaging.class" beforeDir="false" afterPath="$PROJECT_DIR$/target/classes/customer/CustomerTransactionManaging.class" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/target/classes/main/MainWork.class" beforeDir="false" afterPath="$PROJECT_DIR$/target/classes/main/MainWork.class" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -208,7 +215,8 @@
       <workItem from="1677445691690" duration="8872000" />
       <workItem from="1677491588134" duration="1489000" />
       <workItem from="1677575458808" duration="7181000" />
-      <workItem from="1677615348532" duration="699000" />
+      <workItem from="1677615348532" duration="3063000" />
+      <workItem from="1677664193720" duration="2686000" />
     </task>
     <task id="LOCAL-00001" summary="https://github.com/MilianCode/banking-system.git">
       <created>1674862340659</created>
@@ -413,7 +421,14 @@
       <option name="project" value="LOCAL" />
       <updated>1677600011405</updated>
     </task>
-    <option name="localTasksCounter" value="30" />
+    <task id="LOCAL-00030" summary="https://github.com/MilianCode/banking-system.git">
+      <created>1677616342720</created>
+      <option name="number" value="00030" />
+      <option name="presentableId" value="LOCAL-00030" />
+      <option name="project" value="LOCAL" />
+      <updated>1677616342720</updated>
+    </task>
+    <option name="localTasksCounter" value="31" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
